v1.0 添加了小区信息按顺序查询功能。

+++ src/main/java/com/jzq/bean/HousingEstateInfo.java
@@ -11,9 +11,9 @@ public class HousingEstateInfo {
     private String housingEstate;
     private int price;
     private double size;
-    private double face;
-    private double bedroom;
-    private double subway;
-    private double company;
+    private float face;
+    private float bedroom;
+    private float subway;
+    private float company;
     private int score;
 }
